Fixed eslint dependency issues.
