+++ pages/cart/cart.js
@@ -1,7 +1,34 @@
-import { refreshCartContent } from '../../utils/modal/modal.js';
+// import { refreshCartContent } from '../../utils/modal/modal.js';
 import { formatCurrency } from '../../helpers/helpers.js';
 import { SquareOrderLineItem } from '../../constructors/constructors.js';
 import { loadCSS } from '../../scripts/aem.js';
+import { orderForm } from '../../utils/orderForm/orderForm.js';
+
+// Function to refresh the cart content
+export function refreshCartContent(element) {
+  const cartContent = element.querySelector('.card-wrapper');
+  if (cartContent) cartContent.remove();
+
+  const emptyCartMessage = element.querySelector('.empty-cart-message');
+  if (emptyCartMessage) emptyCartMessage.remove();
+
+  const cartOrderForm = element.querySelector('.cart-order-form');
+  if (cartOrderForm) cartOrderForm.remove();
+
+  const currentCart = getCart();
+  element.append(currentCart);
+
+  // Check if currentCart contains the class `card-wrapper` (cart with items)
+  if (currentCart.classList.contains('card-wrapper')) {
+    // If cart has items, append the order form
+    const cartKey = getLastCartKey();
+    const cartLocalStorageData = getLocalStorageCart();
+    const hasShipping = !!((cartKey === 'shipping' || cartKey === 'merch'));
+
+    const form = orderForm(cartLocalStorageData, hasShipping);
+    element.append(form);
+  }
+}
 
 export const allowedCartPages = Object.freeze([
   'store',
